+++ app/components/install/Tekton/InstallContentsTekton1.tsx
@@ -3,7 +3,7 @@ import React, { useContext } from 'react';
 import { Button } from '@material-ui/core';
 import styles from '../InstallContents1.css';
 import routes from '../../../utils/constants/routes.json';
-import productImage from '../../../../resources/assets/HyperCloud Operator_logo.png';
+import productImage from '../../../../resources/assets/Tekton_logo.png';
 import * as env from '../../../utils/common/env';
 import CONST from '../../../utils/constants/constant';